requirements.txt
streamlit
aiohttp
pandas
python-dotenv